server/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
django-environ==0.10.0
Flask==2.2.2
Flask-HTTPAuth==4.7.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
py-mjpeg==1.0.1
requests==2.28.1
urllib3==1.26.12
Werkzeug==2.2.2
